+++ src/elements/db/ElementQuery.php
@@ -29,6 +29,7 @@ use craft\helpers\ElementHelper;
 use craft\helpers\StringHelper;
 use craft\models\Site;
 use craft\search\SearchQuery;
+use ReflectionProperty;
 use yii\base\ArrayableTrait;
 use yii\base\Exception;
 use yii\base\InvalidArgumentException;
@@ -1644,11 +1645,10 @@ class ElementQuery extends Query implements ElementQueryInterface
      */
     public function criteriaAttributes(): array
     {
-        // By default, include all public, non-static properties that were defined by a sub class, and certain ones in this class
-        $class = new \ReflectionClass($this);
         $names = [];
 
-        foreach ($class->getProperties(\ReflectionProperty::IS_PUBLIC) as $property) {
+        // By default, include all public, non-static properties that were defined by a sub class, and certain ones in this class
+        foreach ((new \ReflectionClass($this))->getProperties(ReflectionProperty::IS_PUBLIC) as $property) {
             if (!$property->isStatic()) {
                 $dec = $property->getDeclaringClass();
                 if (
@@ -1657,6 +1657,18 @@ class ElementQuery extends Query implements ElementQueryInterface
                 ) {
                     $names[] = $property->getName();
                 }
+            }
+        }
+
+        // Add custom field properties
+        /** @var CustomFieldBehavior $behavior */
+        $behavior = $this->getBehavior('customFields');
+        foreach ((new \ReflectionClass($behavior))->getProperties(ReflectionProperty::IS_PUBLIC) as $property) {
+            if (
+                !$property->isStatic() &&
+                !in_array($property->getName(), ['hasMethods', 'owner'], true)
+            ) {
+                $names[] = $property->getName();
             }
         }